Renamed project dir to matatika_notebook and updated setup to py to reflect. Fixes package name collision. Fixed local module imports
